requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
matplotlib
tqdm
scikit-learn
albumentations
segmentation-models-pytorch
pillow
